Polishing + Attempt at tying together
Lui's commit does not compile and is lacking sound. This made mine crash
when I merged.

To make it "work", I commented out in resource.js the attempt to get all
sound files (as they were not included in the commit), and made main.js
go straight to game so it would run.