Server now sends JSON files, stores them and sorts
C# Server now saves all files with a ID, Score and JSON file. It sorts
the data when a client wants the results and will only send up to a
given amount (the "Top 50" or "Top 100" amount). Communicates properly,
fixed some bugs but have it working.

All that is left for the server (other then stress testing) is to
encrypt the data being sent then decode it server-side.