Server creates Achievements
When sending a score to the server, the server will now send back to the
client a JSONP file of a function which contains as a parameter a JSON
array of Achievements objects, each with contain a Title, Details and
Image path for that achievement. These achievements are based on what
that specific score qualified for, IE, "Top 10", "Top 25", "Speed
Racer", etc.